requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
django-filter==1.1.0
django-rest-swagger==2.1.2
django==1.11.10
djangorestframework==3.7.7
idna==2.6                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
markupsafe==1.0           # via jinja2
openapi-codec==1.3.2      # via django-rest-swagger
pytz==2018.3              # via django
requests==2.18.4          # via coreapi
simplejson==3.13.2        # via django-rest-swagger
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
